rename package and keyword validation
