+++ client/src/products/about/AboutTabKeywords.tsx
@@ -6,7 +6,7 @@ import "./about-tab-keywords.scss";
 import { SeriesRegistrationDTO } from "utils/types/response-types";
 import { updateSeriesKeywords } from "api/SeriesApi";
 import { useErrorStore } from "utils/store/useErrorStore";
-import { TEST_Combobox } from "products/about/comboboxtest/combobox";
+import { KeywordInput } from "products/about/keyword-input/combobox";
 
 interface Props {
   series: SeriesRegistrationDTO;
@@ -19,7 +19,6 @@ export const AboutTabKeywords = ({ series, isAdmin, mutateSeries, isEditable }: 
   const keywords = series.seriesData.attributes.keywords;
   const [showEditKeywordsMode, setShowEditKeywordsMode] = useState(false);
   const [keywordFormatError, setKeywordFormatError] = useState<string | undefined>(undefined);
-  const [inputValue] = useState("");
   const [updatedKeywords, setUpdatedKeywords] = useState<string[]>(keywords ? keywords : []);
 
   const { setGlobalError } = useErrorStore();
@@ -27,20 +26,29 @@ export const AboutTabKeywords = ({ series, isAdmin, mutateSeries, isEditable }: 
   const validKeywordLetters = new RegExp(/^[A-Za-zÀ-ÖØ-öø-ÿ0-9_\s]*$/);
 
   const handleSaveKeywords = () => {
+    if (updatedKeywords.length <= 3 && updatedKeywords.every((keyword) => allowedCharacters(keyword))) {
+      updateSeriesKeywords(series!.id, updatedKeywords, isAdmin)
+        .then(() => mutateSeries())
+        .catch((error) => {
+          setGlobalError(error.status, error.message);
+        });
+      setShowEditKeywordsMode(false);
+    }
+  };
+
+  const allowedCharacters = (keyword: string) => isValidKeyword(keyword) && validKeywordLetters.test(keyword);
+
+  const validKeyword = (keyword: string) => {
     setKeywordFormatError(undefined);
-    if (updatedKeywords.length <= 3) {
-      if (
-        updatedKeywords.map((keyword) => isValidKeyword(keyword) && validKeywordLetters.test(keyword)).every(Boolean)
-      ) {
-        updateSeriesKeywords(series!.id, updatedKeywords, isAdmin)
-          .then(() => mutateSeries())
-          .catch((error) => {
-            setGlobalError(error.status, error.message);
-          });
-        setShowEditKeywordsMode(false);
-      } else
-        setKeywordFormatError("Blir ikke lagret, nøkkelord kan bare inneholde norske bokstaver, tall og underscore");
-    } else setKeywordFormatError("Du kan maksimalt velge 3 nøkkelord");
+    if (updatedKeywords.length >= 3) {
+      setKeywordFormatError("Du kan maksimalt velge 3 nøkkelord");
+      return false;
+    }
+    if (!allowedCharacters(keyword)) {
+      setKeywordFormatError("Nøkkelord kan bare inneholde norske bokstaver, tall og underscore");
+      return false;
+    }
+    return true;
   };
 
   return (
@@ -49,19 +57,6 @@ export const AboutTabKeywords = ({ series, isAdmin, mutateSeries, isEditable }: 
         <Heading level="2" size="xsmall">
           Nøkkelord
         </Heading>
-        <TEST_Combobox
-          label={"test"}
-          options={[]}
-          clearButton={true}
-          selectedOptions={updatedKeywords || []}
-          maxSelected={{ limit: 3 }}
-          onToggleSelected={(option: string, isSelected: boolean) =>
-            isSelected && isValidKeyword(inputValue) && validKeywordLetters.test(inputValue)
-              ? setUpdatedKeywords([...updatedKeywords, option])
-              : setUpdatedKeywords(updatedKeywords.filter((o) => o !== option))
-          }
-          error={keywordFormatError}
-        />
         <BodyShort>Stikkord som kan søkes på i søket til FinnHjelpemiddel</BodyShort>
         {!showEditKeywordsMode && (
           <>
@@ -101,20 +96,14 @@ export const AboutTabKeywords = ({ series, isAdmin, mutateSeries, isEditable }: 
 
         {showEditKeywordsMode && (
           <>
-            <UNSAFE_Combobox
-              className="keyword-box"
-              id="keywords"
-              label="Legg til nøkkelord"
-              allowNewValues={true}
-              isMultiSelect={true}
-              clearButton={true}
+            <KeywordInput
+              label={"Nøkkelord input"}
+              hideLabel
               options={[]}
-              selectedOptions={updatedKeywords || []}
+              selectedOptions={updatedKeywords}
               maxSelected={{ limit: 3 }}
-              shouldShowSelectedOptions={true}
-              shouldAutocomplete={true}
               onToggleSelected={(option: string, isSelected: boolean) =>
-                isSelected && isValidKeyword(inputValue) && validKeywordLetters.test(inputValue)
+                isSelected && validKeyword(option)
                   ? setUpdatedKeywords([...updatedKeywords, option])
                   : setUpdatedKeywords(updatedKeywords.filter((o) => o !== option))
               }
