--- client/src/products/about/comboboxtest/combobox/index.ts
@@ -1,2 +1,0 @@
-export { default as TEST_Combobox } from "./ComboboxProvider";
-export { type ComboboxProps } from "./types";
