+++ client/src/products/about/keyword-input/combobox/index.ts
@@ -1,0 +1,2 @@
+export { default as KeywordInput } from "./ComboboxProvider";
+export { type ComboboxProps } from "./types";
